synq: switch Python bindings to in-process FFI transport
The Python package spoke JSON-RPC to the `syntaqlite` CLI over a
subprocess. That can't work under Pyodide/WebAssembly (no process
spawning) and is the blocker to publishing a WASM wheel to PyPI
(PEP 783 / Pyodide 314).

Replace it with an in-process transport that calls the serve handler
directly over a C ABI, and make it the *sole* transport: a parser that
can crash the host is a bug to fix at the source, not to sandbox behind
a subprocess, so a fallback earns nothing.

- syntaqlite/src/serve/: extract the per-request serve handler out of the
  CLI binary into a reusable module shared by the CLI stdio loop and the
  FFI (no logic duplicated).
- syntaqlite-ffi: new cdylib+staticlib crate exposing a tiny C ABI
  (serve_create_sqlite/_create_dialect/_call/_free/_destroy). Every
  extern "C" body is wrapped in catch_unwind so a panic becomes an
  {"ok":false,...} envelope instead of unwinding across FFI.
- python/syntaqlite/__init__.py: a single _InProcessTransport (ctypes
  over the cdylib; ctypes releases the GIL on each call). No subprocess,
  no transport selection, no `binary` kwarg.
- setup.py / release.yml: build and bundle the cdylib into syntaqlite/lib/.
- python-api suite self-locates the cdylib; docs updated.

The CLI's `serve json` command stays (it shares the same handler and is
still covered by the serve suite).

Verified: cargo check/clippy --all-features clean; serve + python-api
suites green (49/49) through the in-process transport.
