+++ syntaqlite/Cargo.toml
@@ -1,5 +1,3 @@
-# @generated by syntaqlite-buildtools — DO NOT EDIT
-
 [package]
 name = "syntaqlite"
 version = "0.5.10"
@@ -11,7 +9,7 @@ homepage.workspace = true
 readme = "README.md"
 
 [lib]
-crate-type = ["rlib", "staticlib"]
+crate-type = ["rlib", "staticlib", "cdylib"]
 
 [features]
 default = ["sqlite", "fmt", "analysis"]
@@ -20,6 +18,7 @@ analysis = ["syntaqlite-syntax/analysis"]       # schema analysis (unknown refs,
 experimental-embedded = ["analysis"]    # [experimental] embedded SQL extraction from host languages (Python f-strings, etc.)
 dynload = ["dep:libloading"]             # dynamic dialect loading from .so/.dylib/.dll
 lsp = ["fmt", "analysis", "sqlite", "dep:glob", "dep:lsp-server", "dep:lsp-types", "dep:serde_json"]  # LSP analysis host + protocol server
+rpc = ["fmt", "analysis", "serde-json"]  # transport-agnostic JSON-RPC dispatch (CLI serve loop + C API)
 sqlite = ["syntaqlite-syntax/sqlite"]  # SQLite dialect (grammar, tokens, cflags, functions)
 serde = ["dep:serde", "syntaqlite-syntax/serde"]         # serde::Serialize/Deserialize impls for diagnostics and AST nodes
 serde-json = ["serde", "dep:serde_json", "syntaqlite-syntax/serde-json"]  # JSON convenience helpers (dump_json, from_json)
